+++ my-stocks/src/main/java/dev/hgsaldanha/repository/BrokerRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import dev.hgsaldanha.model.Broker;
 
-public interface BrokerRepository extends JpaRepository<Broker, Long> {
+public interface BrokerRepository extends JpaRepository<Broker, Integer> {
 
 }
